.history/requirements_20251124094618.txt
gymnasium
matplotlib
numpy
pandas
pygame
sklearnpip install scikit-learn

threading
time
torch
torchvision
